Prevent crash when trying to export all decks
The crash was caused by the warning dialog, itself!  The issue was
that export is now run in a background thread, but GUI operations from
a background thread cause a crash.

AnkiModalNotifier is used in `anki_exporter_wrapper.py` and
`github_importer.py`.  The former is in a background thread, the
latter in the main thread.  In both cases, error messages are now
correctly displayed, without crashing.

Tested on Anki 2.1.40 and 2.1.43.