Mock mw in test/export/note_sorter_spec.py
`mw` is None, by default (see `qt/aqt/__init__.py`) and is only
properly set on init.  Since our tests don't carry out the full Anki
initialisation (and in the case of this test, probably shouldn't —
it's supposed to be a light unit test), `mw` is None.

I don't think that mocking `mw` here is "wrong" — here we're not
necessarily interested in testing full integration with Anki (just
that we're roughly API-compatible and that the logic is correct).

With this change, the export tests all pass:

pipenv run mamba  ./test/export/