Add test for succesful repeated "import" (clone and pull) from git
Diagnose #138 which occurred on Windows.  (A lock for the git pack
file is often held by a previous clone/pull and not released, so the
next pull fails (`PermissionError`).)

The monkey-patching of `get_repo_local_path` is ugly, but far less
complex than partially mocking ConfigSettings.  Alternatively, we
could have modified `get_repo_local_path` to have an extra, optional
argument (say `full_snapshot_path` that'd override the config call),
but modifying the code to fit the test is also ugly.