Fix github workflow
See: dulwich issue 1093.

Note that with our current version of dulwich (0.20.24) this seems to
build dulwich with C bindings.  For testing this is suboptimal but not
critical (it's unlikely that dulwich's behaviour differs depending on
whether the C bindings are used or not).  (It will be critical for
building crowd_anki, though.)