Force protobuf python implementation
It seems that Anki 2.1.35 (the last version compatible with python3.7)
is not actually compatible with the latest protobuf (possibly released
after Anki 2.1.35).

See:
https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates

Both anki and protobuf are dev dependencies so this incompatibility
only affects testing.

We will be able to remove this workaround by upgrading to
python3.9 (compatible with latest Anki).