Allow testing building crowd_anki/dist when Pipfile.lock is changed
Fix #182.

The uploaded artifact is currently unused.

We check for changes to both Pipfile and Pipfile.lock to avoid stale a
Pipfile.lock.

This will result in some false positive runs of this workflow, when
we've, say, added a new script to Pipfile, without changing
dependencies.

We will also have false positives when only dev dependencies have been
updated.