Force building pyyaml without libyaml
pyyaml detects the presence of the libyaml library, and builds with
the libyaml extension if it's present.  In principle, there's a
--without-libyaml flag, to ensure building without libyaml even if the
library is present, but it doesn't work.  (Also, there is an issue
with passing on `--without-libyaml` via pip to setup.py.)
PYYAML_FORCE_LIBYAML=0 is not documented, but works.

Another workaround would be to ensure that libyaml is not present, but
that's brittle, since then the build system starts depending on global
configuration (presence/absence) of system libraries.