Replace Qt5 attributes with Qt6 ones
This allows CrowdAnki on Anki 23.10rc1 to run without
ENABLE_QT5_COMPAT=1.

I believe that compatibility with Qt5 and Anki 2.1.50->2.1.66 is
preserved.

The Qt6 version of the code is in the try clause, while the Qt5 one is
in the except.  Hopefully, this will make clean up (once Qt5 is
dropped by Anki) simple.

There are still some Qt-unrelated deprecation warning with 23.10rc1,
but it's probably cleanest to deal with them separately, since I'm not
sure how easy it will be to make the relevant changes
backwards-compatible.