Update to python3.10
Fix #225.

Anki currently uses 3.9.  I think that being one version ahead
shouldn't be that much of an issue if our dependencies don't start
jumping onto new python features.  It has the significant advantage of
avoiding update churn — this upgrade turned out completely painless,
but the previous attempt, ~two years ago, was a sufficient nightmare
that I decided to postpone it.  (I don't quite remember, but the main
issues were: 1.  conflicting transitive dependencies, 2. a new version
of one of our dependencies no longer working in the pure python
version (and the old version being incompatible with something
else??), 3. pip weirdness.)  Given that python3.9 is EOL in 2025-10,
Anki will probably be updating to 3.10 (or later) soon anyway.

(The intial reason was that 3.10 is the default on my system (though I
also have most other current versions installed), but I think the
justification for it over 3.9 or 3.11 is sound.)