Pin PyQt5 et al. for windows compat
Pipfile.lock is still generated on Linux.

(An alternative was generating the lockfile on windows, which would
mean that pywin would be installed (on windows) and wouldn't have to
be mocked.  Unfortunately, this would be annoying in the future, since
all future Pipfile.lock files would also have to be generated on
Windows (pipenv happily ignores windows-only dependencies when
pipenv-installing on linux, but it doesn't preserve them when
upgrading the lockfile).)