Use setup-python python on Windows, in GitHub actions
The underlying issue is that the `windows-latest` runner in GitHub
actions already has a 3.10 version of python installed, which is
32-bit.  (However, the in-built 3.9 version is 64-bit.)

(This pre-installed version is at
C:\hostedtoolcache\windows\Python\3.10.11\x86\python.exe .  The
version installed by setup-python is at
C:\hostedtoolcache\windows\Python\3.10.11\x64\python.exe .)

Unfortunately, when one doesn't specify the python location with
`--python`, then pipenv uses the pre-installed 32-bit version.

This is an issue because the `anki` package only has
arch/system-specific wheels (after 2.1.35, which was the last
having (only) a generic wheel), and the only windows wheel is for
`win_amd64` (i.e. 64-bit), while the 32-bit version of python wants a
`win32` wheel.  Hence, with the (as of now) current anki==25.2.5 in
our Pipfile.lock, `pipenv install --dev` would fail in the Windows
GitHub mamba action.

The pipenv install failure was confirmed to be present with the
windows-latest (as of 2025-05-16), windows-2022 and windows-2019 runners.

Forcing pipenv to use the version/location of python installed with
setup-python solves the issue.